+++ frontend/src/pages/Help.vue
@@ -2275,7 +2275,7 @@ OluFxewsEO0QNDrfFb+0gnjYlnGqOFcZjUMXbDdY5oLSPtXohynuTK1qyQ==
 </div>
 
 <div class="text-center small text-dim" v-pre>
-Automatically generated from Statping's Wiki on 2020-09-02 02:46:04.864615 &#43;0000 UTC
+Automatically generated from Statping's Wiki on 2020-09-09 00:56:11.888825 &#43;0000 UTC
 </div>
 
 </div>