+++ CHANGELOG.md
@@ -1,9 +1,13 @@
+# 0.90.66 (09-08-2020)
+- Added Import and Export views in Dashboard
+
 # 0.90.65 (09-01-2020)
 - Fixed issue with dashboard not logging in (notifier panic)
 - Modified static email templates to github.com/statping/emails
 - Modified Regenerate API function to keep API_SECRET env
 - Added DEMO_MODE env variable, if true, 'admin' cannot be deleted
 - Modified Service sparklines on Dashboard
+- Added modal popup for UI deletes/edits
 
 # 0.90.64 (08-18-2020)
 - Modified max-width for container to 1012px, larger UI